% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_thaliana.R
\docType{data}
\name{A.thaliana}
\alias{A.thaliana}
\title{Genetic and geographic data for Arabidopsis thaliana}
\format{A list with 3 arguments: genotype, coord, chrpos}
\usage{
data(A.thaliana)
}
\description{
A dataset containing SNP frequency and geographic data for 170 plant accessions.
The variables are as follows:
}
\details{
\itemize{
  \item genotype binary (0 or 1) SNP frequency for 170 accessions (26943 SNPs).
  \item coord longitude and latitude of individuals
  \item chrpos genetic map, chromosome, position of each SNP
}
}
\keyword{datasets}
