% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_pheno.R
\name{create_factor}
\alias{create_factor}
\title{Generates confounders.}
\usage{
create_factor(genotype, K)
}
\arguments{
\item{genotype}{an n by p matrix of genotypes where n is the number of individuals (rows) and p is the
number of SNPs (columns).}

\item{K}{an integer for the number of confounders.}
}
\value{
A \code{list} containing K factors, an estimate of residual error, and an estimate of
the order of effects.

factors a n by K matrix of confounders corresponding to the first K unscaled principal
components of the genotype matrix.

sigma standard deviation for residual errors.

base order of magnitude for baseline effects.
}
\description{
This function creates confounding factors based on genotypes. Those factors may be included
in the simulation of phenotypes.
}
\examples{
library(naturalgwas)

## Load A. thaliana example
data(A.thaliana)
confounder <- create_factor(A.thaliana$genotype, K = 20)
plot(confounder$factors)
cat("Residual error = ", confounder$sigma, "\\n")
cat("Base effect = ", confounder$base, "\\n")
}
\seealso{
\code{\link{simu_pheno}}
}
\author{
Olivier François
}
