% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_pheno.R
\name{get_climate}
\alias{get_climate}
\title{Evaluates environment from 'worldclim'}
\usage{
get_climate(coord)
}
\arguments{
\item{coord}{a matrix of geographic coordinates written as (longitude, latitude).}
}
\value{
the first principal component of 19 bioclimatic variables from 'worldclim'.
}
\description{
This function downloads the 'wordlclim' bioclimatic database, extract the data for the specified geographic coordinates,
and summarizes them with their first principal component.
}
\examples{
library(naturalgwas)

## Load example
data(A.thaliana)
env <- get_climate(A.thaliana$coord)
summary(env)
}
\seealso{
\code{\link{simu_pheno}}
}
\author{
Olivier François
}
