% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_pheno.R
\name{simu_pheno}
\alias{simu_pheno}
\title{Simulate phenotypes based on genotypes.}
\usage{
simu_pheno(genotype, confounder, environment, ref.set, ncausal, effect.size,
  gxe)
}
\arguments{
\item{genotype}{an n by p matrix of SNP genotypes where n is the number of individuals (rows) and p is the
number of SNPs (columns).}

\item{confounder}{an object of class "confounder" created with \code{\link{create_factor}}.}

\item{environment}{an environmental vector created \code{\link{get_climate}}.}

\item{ref.set}{a set of weakly linked reference SNPs created with \code{\link{create_refset}}.}

\item{ncausal}{number of causal variants in the simulation.}

\item{effect.size}{effect size for causal variants in the simulation.}

\item{gxe}{intensity of gene by environment interaction in the simulation.}
}
\value{
A subset of 'ncausal' causal variants and a vector of phenotypes.

causal.set  subset of 'ncausal' causal variants.

phenotype  vector of phenotypes.
}
\description{
This function samples causal variants from a reference set of variants, and simulates phenotypes based on
genotypes, enviromnental variables, and confounding factors.
}
\examples{
library(naturalgwas)

## Load A. thaliana example
data(A.thaliana)
env <- get_climate(A.thaliana$coord)
confounder <- create_factor(A.thaliana$genotype, K = 20)
ref.set <- create_refset(A.thaliana$chrpos, window = 101)
sim.ph <- simu_pheno(A.thaliana$genotype, confounder, env, ref.set,
ncausal = 20, effect.size = 100, gxe = 1)
}
\seealso{
\code{\link{create_refset}} \code{\link{create_factor}} \code{\link{get_climate}}
}
\author{
Olivier François
}
