% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_thaliana.R
\docType{data}
\name{A.thaliana}
\alias{A.thaliana}
\title{Genetic and geographic data for Arabidopsis thaliana}
\format{A list with 4 arguments: genotype, coord, chrpos, f16, ecotype.id}
\usage{
data(A.thaliana)
}
\description{
A dataset containing SNP frequency and geographic data for 162 European
plant accessions from Atwell et al. Science 2010.
The variables are as follows:
}
\details{
\itemize{
  \item genotype: binary (0 or 1) SNP frequency for 162 European accessions (53859 SNPs).
  \item coordinates: longitude and latitude of individuals.
  \item chrpos: genetic map, chromosome (Chr 5) and position for each SNP.
  \item f16: days to flowering trait (FT16) from Atwell et al. 2010. Plants
  were checked bi-weekly for presence of first buds, and the average flowering time of
  4 plants of the same accession were collected.
 \item ecotype.id: ecotype.id, native name, site and region for 162 European accessions.
}
}
\keyword{datasets}
